+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Set `blockOwnerDeletion` to `true` in the owner references in Strimzi managed resources.
   Deleting the Strimzi custom resources will now by default wait for the deletion of all the owned Kubernetes resources.
+* New fields `.spec.groupId`, `.spec.configStorageTopic`, `.spec.offsetStorageTopic`, and `.spec.statusStorageTopic` in the `KafkaConnect` custom resource for configuring Connect's group ID and internal topics.
 
 ### Major changes, deprecations, and removals
 
@@ -16,6 +17,8 @@
   To use the Keycloak authorizer, you can use the `type: custom` authorization.
   The Strimzi OAuth library with the Keycloak authorizer continues to be packaged with the Strimzi container images.
   Follow the documentation for the examples and migration details.
+* The `group.id`, `config.storage.topic`, `offset.storage.topic`, and `status.storage.topic` fields in `.spec.config` section of the `KafkaConnect` resource are deprecated and will not be supported in the `v1` CRD API.
+  Please use the new `.spec.groupId`, `.spec.configStorageTopic`, `.spec.offsetStorageTopic`, and `.spec.statusStorageTopic` fields instead.
 
 ## 0.48.0